# openapi.unified.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TicketRemaster Unified API",
    "version": "2026-03-29",
    "description": "Unified OpenAPI document for the current TicketRemaster gateway-facing orchestrator API plus the external OutSystems credit service."
  },
  "servers": [
    {
      "url": "https://ticketremasterapi.hong-yi.me",
      "description": "Production gateway"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local gateway"
    }
  ],
...
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login and receive a JWT",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logged in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Get current profile",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          }
        }
      }
    },
...
    "/credits": {
      "post": {
        "tags": ["External Credit Service"],
        "summary": "Create or initialize credit record",
        "servers": [
          {
            "url": "https://personal-sdxnmlx3.outsystemscloud.com/CreditService/rest/CreditAPI",
            "description": "OutSystems credit service"
          }
        ],
        "security": [
          {
            "OutSystemsApiKey": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExternalCreateCreditRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Credit record created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalCreditResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "GatewayApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "apikey"
      },
      "OutSystemsApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-KEY"
      }
    },
    "responses": {
      "Error400": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Error401": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
...
      "Error409": {
        "description": "Conflict",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
...
    },
    "schemas": {
      "DataEnvelope": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "status": {
                "type": "integer"
              },
              "traceId": {
                "type": "string"
              }
            },
            "required": ["code", "message"]
          }
        },
        "required": ["error"]
      },
      "RegisterRequest": {
        "type": "object",
        "required": ["email", "password", "phoneNumber"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "venueId": {
            "type": "string"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        }
      },
...
      "ExternalCreateCreditRequest": {
        "type": "object",
        "required": ["userId", "creditBalance"],
        "properties": {
          "userId": {
            "type": "string"
          },
          "creditBalance": {
            "type": "number"
          }
        }
      },
...
      "ExternalCreditResponse": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "creditBalance": {
            "type": "number"
          },
          "lastToppedUpAt": {
            "type": "string",
            "format": "date-time"
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      }
    }
  }
}